PRJNA517180

A Myt1 family transcription factor defines neuronal fate by repressing non-neuronal genes.

Caenorhabditis elegans.

RNA-seq data from PRJNA51780. In this experiment, the importance of ztf-11 for neurogenesis was evaluated in order to infer the role of Myt1, the Homo sapiens homologous gene. Myt1 and ztf-11 are zinc-finger transcription factors that are thought to be important in the reprogrammation from fibroblast to neuron in vitro. There are two groups: the control group consists of individuals with a control vector (L4440), which does not affect physiology. On the other hand, the case group consists of knockdown mutants of ztf-11 gene via RNAi.

Format

Large RangedSummarizedExperiment

Source

https://www.ncbi.nlm.nih.gov/bioproject/PRJNA517180

Examples

data(PRJNA51780,package="enriquepresa")

#> Warning: data set ‘PRJNA51780’ not found